Create NphOrderForm class and extend in order collect & sample finalize form
