+++ src/Form/Nph/NphSampleFinalize.php
@@ -1,0 +1,38 @@
+<?php
+
+namespace App\Form\Nph;
+
+use Symfony\Component\Form\FormBuilderInterface;
+use Symfony\Component\Form\Extension\Core\Type;
+use Symfony\Component\OptionsResolver\OptionsResolver;
+use Symfony\Component\Validator\Constraints;
+
+class NphSampleFinalize extends NphOrderForm
+{
+    public function buildForm(FormBuilderInterface $builder, array $options)
+    {
+        $sample = $options['sample'];
+        $orderType = $options['orderType'];
+
+        $this->addCollectedTimeAndNoteFields($builder, $options, $sample);
+
+        if ($orderType === 'urine') {
+            $this->addUrineMetadataFields($builder);
+        }
+
+        if ($orderType === 'stool') {
+            $this->addStoolMetadataFields($builder);
+        }
+
+        return $builder->getForm();
+    }
+
+    public function configureOptions(OptionsResolver $resolver)
+    {
+        $resolver->setDefaults([
+            'sample' => null,
+            'orderType' => null,
+            'timeZone' => null
+        ]);
+    }
+}
